Added Audit table for the price change and changed the price datatype from INT to FLOAT

+++ SQL/Table Creation/PubOrderDetail.sql
@@ -3,7 +3,7 @@ CREATE TABLE Pub_Order_Detail(
     OrderID INT NOT NULL,
     ProdID INT NOT NULL,
     Qty INT NOT NULL,
-    Price INT NOT NULL,
+    Price FLOAT NOT NULL,
     PRIMARY KEY (OrderDetailsID),
     FOREIGN KEY (OrderID) REFERENCES pub_order(OrderID),
     FOREIGN KEY (ProdID) REFERENCES pub_product(ProdID),
